requirements.txt
crewai
crewai-tools
requests
beautifulsoup4
pandas
python-dotenv
streamlit
LiteLLM
